+++ pywatts/summaries/__init__.py
@@ -4,3 +4,4 @@ from pywatts.summaries.mase_summary import MASE
 from pywatts.summaries.max_summary import MaxErr
 from pywatts.summaries.min_summary import MinErr
 from pywatts.summaries.rmse_summary import RMSE
+from pywatts.summaries.tsne_visualisation import TSNESummary